DEV - Pass image to popUp
