+++ 4-Object_Detection/YOLOV3/core/utils.py
@@ -3,6 +3,10 @@ import random
 import colorsys
 import numpy as np
 from core.config import cfg
+from PyQt5 import QtGui
+from PyQt5.QtWidgets import QWidget, QApplication, QLabel, QVBoxLayout
+from PyQt5.QtGui import QPixmap
+from PyQt5.QtCore import pyqtSignal, pyqtSlot, Qt, QThread
 # sys.path.append(".")
 from object_classes.ImageDetected import ImageDetected
 
@@ -236,6 +240,14 @@ def postprocess_boxes(pred_bbox, org_img_shape, input_size, score_threshold):
 
     return np.concatenate([coors, scores[:, np.newaxis], classes[:, np.newaxis]], axis=-1)
 
-
+def convert_cv_qt(cv_img):
+    
+    """Convert from an opencv image to QPixmap"""
+    rgb_image = cv2.cvtColor(cv_img, cv2.COLOR_BGR2RGB)
+    h, w, ch = rgb_image.shape
+    bytes_per_line = ch * w
+    convert_to_Qt_format = QtGui.QImage(rgb_image.data, w, h, bytes_per_line, QtGui.QImage.Format_RGB888)
+    p = convert_to_Qt_format.scaled(661, 521, Qt.KeepAspectRatio)
+    return QPixmap.fromImage(p)
